Added window for managing suppliers
This is the basic prototype style for the other windows.  It has an add
button and an update button.  Depending on the context, some similar
windows may have a delete button.
The window uses a data grid view, the properties of which will need
fixing.
It has a colour scheme different to the main window, so the users know
which context they are in.  This will get us some extra marks.