Window Behaviour Fixed for most UI
With the exception of the report windows, all windows now open and close
without causing errors.  The technique used is one recommended by the
lecturer, and saves on CPU when opening windows.  The program creates
all windows that are needed, hides them until they are needed, in which
case, it shows the window, (Win.Show() instead of Window Win = new
Window () )  and when the window is not needed, it is hidden,
(Win.Hide() instead of Window Win = new Window() ).  It is a much neater
and elegant solution.